Added more general fixes"

+++ src/main/java/com/controletcc/facade/ModeloAvaliacaoFacade.java
@@ -8,6 +8,8 @@ import com.controletcc.model.entity.ModeloAvaliacao;
 import com.controletcc.repository.projection.ModeloAvaliacaoProjection;
 import com.controletcc.service.ModeloAvaliacaoService;
 import com.controletcc.service.ProfessorService;
+import com.controletcc.service.ProjetoTccNotaService;
+import com.controletcc.service.ProjetoTccService;
 import com.controletcc.util.ModelMapperUtil;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -23,6 +25,10 @@ public class ModeloAvaliacaoFacade {
     private final ModeloAvaliacaoService modeloAvaliacaoService;
 
     private final ProfessorService professorService;
+
+    private final ProjetoTccNotaService projetoTccNotaService;
+
+    private final ProjetoTccService projetoTccService;
 
     public ModeloAvaliacaoDTO getById(Long id) {
         var modeloAvaliacao = modeloAvaliacaoService.getById(id);
@@ -43,10 +49,14 @@ public class ModeloAvaliacaoFacade {
     public ModeloAvaliacaoDTO update(ModeloAvaliacaoDTO modeloAvaliacaoDTO) throws Exception {
         var modeloAvaliacao = ModelMapperUtil.map(modeloAvaliacaoDTO, ModeloAvaliacao.class);
         modeloAvaliacao = modeloAvaliacaoService.update(modeloAvaliacao.getId(), modeloAvaliacao);
+        projetoTccNotaService.updateByModelo(modeloAvaliacao);
         return ModelMapperUtil.map(modeloAvaliacao, ModeloAvaliacaoDTO.class);
     }
 
     public void deleteLogic(Long id) throws Exception {
+        if (projetoTccService.existsEmAvaliacaoByModeloAvaliacao(id)) {
+            throw new BusinessException("Não é possível excluir este modelo, pois há projetos de TCC em fase de avaliação que o utilizam.");
+        }
         modeloAvaliacaoService.deleteLogic(id);
     }
